Switching back to js-based output

+++ app/controllers/renderers_controller.rb
@@ -8,6 +8,8 @@ class RenderersController < ApplicationController
 
   def render_scad
 
+    output_format = "THREE.Geometry"
+
     # SCAD => STL
     # ==========================================
 
@@ -26,7 +28,7 @@ class RenderersController < ApplicationController
 
       puts  "\n\n" + "-"*30 + "\n\n" + IO.read(scad_file.path) + "\n\n"
 
-      api_response = extopenscad File.new(scad_file)
+      api_response = extopenscad File.new(scad_file), (output_format == "THREE.Geometry" ? "stl" : "obj")
       api_json = JSON.parse api_response.to_s
 
       puts api_json["message"].inspect + "\n\n" + "-"*30
@@ -43,25 +45,41 @@ class RenderersController < ApplicationController
     # ==========================================
 
     if errored == false
-      #puts "stl -> obj"
-      #obj = SCAD.convert_format(api_json["output"], 'stl', 'obj')[:output]
-      puts "obj -> utf"
-      #obj = IO.read File.join Rails.root, "tools", "ben_00.obj"
-      obj = api_json["output"]
+      if output_format == "utf-8"
 
-      obj_file = File.new(File.join(Rails.root, "test.obj"), "w")
-      obj_file.write obj
-      obj_file.close
+        puts "obj -> utf"
+        #obj = IO.read File.join Rails.root, "tools", "ben_00.obj"
+        obj = api_json["output"]
+
+        obj_file = File.new(File.join(Rails.root, "test.obj"), "w")
+        obj_file.write obj
+        obj_file.close
       
-      #puts obj
-      utf_files = SCAD.obj_to_utf(obj)[:output]
-      utf = utf_files.collect do |f|
-        puts f.path
-        s = IO.read(f)
-        File.unlink(f.path)
-        s
+        #puts obj
+        utf_files = SCAD.obj_to_utf(obj)[:output]
+        utf = utf_files.collect do |f|
+          puts f.path
+          s = IO.read(f)
+          File.unlink(f.path)
+          s
+        end
+        output = utf
+        #puts utf.inspect
+
+      elsif output_format == "THREE.Geometry"
+
+        stl = Tempfile.new(["stl", ".stl"], File.join(Rails.root, "tmp"))
+        begin
+          stl.write api_json["output"]
+          stl.close(false)
+
+          js = ThreeJSMonkey.compile_js(File.new stl.path )
+          output = js
+        ensure
+          stl.close(true)
+        end
+
       end
-      #puts utf.inspect
 
       #js = ThreeJSMonkey.compile_js(File.new stl.path )
     end
@@ -73,20 +91,20 @@ class RenderersController < ApplicationController
     data = {
       :stdout => api_json["message"],
       :stderr => errored ? (api_json["message"].present? ? api_json["message"] : "Uh oh, that went horribly wrong. Try again?") : "",
-      :data => defined?(utf) ? utf : "",
-      :format => "utf-8"
+      :data => defined?(output) ? output : "",
+      :format => output_format
     }
     render :json => data
     #response.headers["Content-Type"] = "json"
   end
 
 
-  def extopenscad(f)
+  def extopenscad(f, format)
     #implicitcad_api_server = Rails.env.development? ? "localhost:3000" : "23.21.177.106:3000"
-    implicitcad_api_server = Rails.env.development? ? "172.16.42.4:3000" : "23.21.177.106:3000"
-    #implicitcad_api_server = "23.21.177.106:3000"
+    #implicitcad_api_server = Rails.env.development? ? "172.16.42.4:3000" : "23.21.177.106:3000"
+    implicitcad_api_server = "23.21.177.106:3000"
 
-    RestClient.post("#{implicitcad_api_server}/v1/render", :file => f, :format => "obj") {|response, request, result| response }
+    RestClient.post("#{implicitcad_api_server}/v1/render", :file => f, :format => format) {|response, request, result| response }
   end
 
   add_method_tracer :extopenscad, 'Custom/RenderController/extopenscad', :metric => true
